Move vect photon to mainline code * all code will be vectorized, so mainlining it * photon.py -> old/photon.py * vect/vphoton.py -> photon.py * added vect.py to hold common constants and eventually helper functions * removed vect subfolder
